Rename RemoteConnect.vue to Login.vue
